Migrated Gulp related build tasks to its own module.
